+++ src/parser.rs
@@ -808,13 +808,14 @@ fn parse_switch(
         }
     }
 
-    let mut table = BTreeMap::<u64, Box<StmtBlock>>::new();
+    let mut table = BTreeMap::<u64, Box<(Option<Expr>, StmtBlock)>>::new();
+    let mut def_pos = Position::NONE;
     let mut def_stmt = None;
 
     loop {
         const MISSING_RBRACE: &str = "to end this switch block";
 
-        let expr = match input.peek().unwrap() {
+        let (expr, condition) = match input.peek().unwrap() {
             (Token::RightBrace, _) => {
                 eat_token(input, Token::RightBrace);
                 break;
@@ -825,12 +826,31 @@ fn parse_switch(
                         .into_err(*pos),
                 )
             }
-            (Token::Underscore, _) if def_stmt.is_none() => {
+            (Token::Underscore, pos) if def_stmt.is_none() => {
+                def_pos = *pos;
                 eat_token(input, Token::Underscore);
-                None
+
+                let (if_clause, if_pos) = match_token(input, Token::If);
+
+                if if_clause {
+                    return Err(PERR::WrongSwitchCaseCondition.into_err(if_pos));
+                }
+
+                (None, None)
             }
             (Token::Underscore, pos) => return Err(PERR::DuplicatedSwitchCase.into_err(*pos)),
-            _ => Some(parse_expr(input, state, lib, settings.level_up())?),
+            _ if def_stmt.is_some() => return Err(PERR::WrongSwitchDefaultCase.into_err(def_pos)),
+
+            _ => {
+                let case_expr = Some(parse_expr(input, state, lib, settings.level_up())?);
+
+                let condition = if match_token(input, Token::If).0 {
+                    Some(parse_expr(input, state, lib, settings.level_up())?)
+                } else {
+                    None
+                };
+                (case_expr, condition)
+            }
         };
 
         let hash = if let Some(expr) = expr {
@@ -868,7 +888,7 @@ fn parse_switch(
         let need_comma = !stmt.is_self_terminated();
 
         def_stmt = if let Some(hash) = hash {
-            table.insert(hash, Box::new(stmt.into()));
+            table.insert(hash, Box::new((condition, stmt.into())));
             None
         } else {
             Some(stmt.into())
@@ -2875,7 +2895,7 @@ fn make_curry_from_externals(
     let mut statements: StaticVec<_> = Default::default();
     statements.extend(externals.into_iter().map(Stmt::Share));
     statements.push(Stmt::Expr(expr));
-    Expr::Stmt(Box::new(StmtBlock { statements, pos }))
+    Expr::Stmt(Box::new(StmtBlock::new(statements, pos)))
 }
 
 /// Parse an anonymous function definition.